requirements.txt
pillow==8.3.2
chardet==3.0.4
charset-normalizer==3.4.3
pandas==1.4.2
pydantic==2.10.6
pydantic_core==2.27.2
qrcode==7.3.1
requests==2.32.4
